consolidate ViewType into types.ts
