+++ frontend/src/models/funnelsModel.ts
@@ -1,8 +1,7 @@
 import { kea } from 'kea'
 import api from 'lib/api'
 import { toParams } from 'lib/utils'
-import { ViewType } from 'scenes/insights/insightLogic'
-import { SavedFunnel } from '~/types'
+import { SavedFunnel, ViewType } from '~/types'
 import { insightHistoryLogic } from 'scenes/insights/InsightHistoryPanel/insightHistoryLogic'
 import { funnelsModelType } from './funnelsModelType'
