+++ frontend/src/scenes/events/EventsTable.tsx
@@ -15,10 +15,9 @@ import relativeTime from 'dayjs/plugin/relativeTime'
 import LocalizedFormat from 'dayjs/plugin/localizedFormat'
 import { TZLabel } from 'lib/components/TimezoneAware'
 import { keyMapping } from 'lib/components/PropertyKeyInfo'
-import { ViewType } from 'scenes/insights/insightLogic'
 import { ResizableColumnType, ResizableTable } from 'lib/components/ResizableTable'
 import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
-import { EventFormattedType } from '~/types'
+import { EventFormattedType, ViewType } from '~/types'
 import { PageHeader } from 'lib/components/PageHeader'
 import { TableConfig } from 'lib/components/ResizableTable'
 import { propertyDefinitionsModel } from '~/models/propertyDefinitionsModel'
