+++ frontend/src/scenes/insights/InsightHistoryPanel/InsightHistoryPanel.tsx
@@ -3,14 +3,13 @@ import { Tabs, Col, Row, Button, Spin } from 'antd'
 import { Loading } from 'lib/utils'
 import { useValues, useActions } from 'kea'
 import { insightHistoryLogic } from './insightHistoryLogic'
-import { DashboardItemType } from '~/types'
+import { DashboardItemType, ViewType } from '~/types'
 import { DashboardItem, DisplayedType, displayMap } from 'scenes/dashboard/DashboardItem'
 import './InsightHistoryPanel.scss'
 import dayjs from 'dayjs'
 import { dashboardItemsModel } from '~/models/dashboardItemsModel'
 import { router } from 'kea-router'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import { ViewType } from '../insightLogic'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 dayjs.extend(relativeTime)
