+++ frontend/src/scenes/insights/InsightTabs/SessionTab.tsx
@@ -2,10 +2,9 @@ import React from 'react'
 import { useValues, useActions } from 'kea'
 import { PropertyFilters } from 'lib/components/PropertyFilters/PropertyFilters'
 import { SessionFilter } from 'lib/components/SessionsFilter'
-import { ViewType } from '../insightLogic'
 import { trendsLogic } from '../../trends/trendsLogic'
 import { ActionFilter } from '../ActionFilter/ActionFilter'
-import { FilterType } from '~/types'
+import { FilterType, ViewType } from '~/types'
 import { Col, Row, Skeleton } from 'antd'
 import { TestAccountFilter } from '../TestAccountFilter'
 import useBreakpoint from 'antd/lib/grid/hooks/useBreakpoint'
