+++ frontend/src/scenes/insights/InsightTabs/TrendTab/TrendTab.tsx
@@ -7,9 +7,8 @@ import { BreakdownFilter } from '../../BreakdownFilter'
 import { CloseButton } from 'lib/components/CloseButton'
 import { InfoCircleOutlined } from '@ant-design/icons'
 import { trendsLogic } from '../../../trends/trendsLogic'
-import { ViewType } from '../../insightLogic'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { FilterType } from '~/types'
+import { FilterType, ViewType } from '~/types'
 import { Formula } from './Formula'
 import { TestAccountFilter } from 'scenes/insights/TestAccountFilter'
 import { preflightLogic } from 'scenes/PreflightCheck/logic'
