+++ frontend/src/scenes/trends/personsModalLogic.ts
@@ -2,9 +2,8 @@ import { kea } from 'kea'
 import api from 'lib/api'
 import { errorToast, toParams } from 'lib/utils'
 import { cleanFunnelParams, funnelLogic } from 'scenes/funnels/funnelLogic'
-import { ViewType } from 'scenes/insights/insightLogic'
 import { cohortLogic } from 'scenes/persons/cohortLogic'
-import { ActionFilter, FilterType } from '~/types'
+import { ActionFilter, FilterType, ViewType } from '~/types'
 import { personsModalLogicType } from './personsModalLogicType'
 import { parsePeopleParams, TrendPeople } from './trendsLogic'
